requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
decorator==4.4.2
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
imageio==2.9.0
importlib-metadata==1.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.1.3
networkx==2.4
numpy==1.19.1
opt-einsum==3.3.0
pandas==1.1.0
Pillow==7.2.0
pkg-resources==0.0.0
plotly==4.9.0
protobuf==3.12.4
pymia==0.2.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
retrying==1.3.3
scikit-image==0.17.2
scipy==1.5.2
SimpleITK==1.2.4
six==1.15.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.3
termcolor==1.1.0
tifffile==2020.7.24
torch==1.6.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0